requirements.txt
flask==3.0.3
requests==2.32.3
flask-cors==5.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.3